Day 02 Commit
Created a way so that the internal classes could call functions
outside of their scope in Meteoroids.gd (i.e. the draw_object()
function). This also lead to an internalization of most process
and draw calls into their respective classes.

Added a buffer boarder to PlayField so now you cannot see
objects when they leave the area.

Added the ability for the player to shoot, and Bullets expire
after a set amount of time.

Added basic Meteor to test spawning them into the level and
destroying them with the player's bullets.